Bot is now able to digest youtube links
There still has to be work done on actually playing the file. Atm it throws because it cant do things related to the voice channel. Have to put in a function in the super class to handle that kind of stuff, in case other modules might want to use the voice channel.