started real work on the webpanel now
Basic authentication via Discord is working, so is the list of manageable servers. Next step will be to dig out all configurable options and, well, make them configurable. This whole thing will replace the command-based system, which will be pushed to a legacy command module for those of you who still want that functionality. Be warned tho that this module WILL NOT be updated as things progress. Maybe someone will find the time and update it alongside but I doubt that.