implement media streaming
Since I recieved a number of requests for this, here is a version of the
bot which uses media streaming instead of downloading media beforehand.
While it is significantly faster than downloading, I initially had
concerns regarding playback stability. With this I want to test if
stability is affected in a meaningful way and if it's not, I might as
well merge this with master.